refactor(config): one usecase file, config access behind a repository
Two structural corrections.

The usecase was spread across five files because the one-type-per-file
rule was applied over the module's own convention. Every other usecase
here is one file plus its test; discovery.go carries seven types
including its own public data types. Config now does the same, at 387
lines against runtime.go's 529.

Config access was an interface and implementation living inside the
usecase package, which also made it the only usecase constructing its
own dependency instead of receiving it from the repository container.
Repository here means app-layer data access, not event-sourced
aggregate — discovery and graph own no aggregate either — so config
belongs with them, and now wires through repos.Config like its siblings.

Behaviour, API shape and swagger output are unchanged.
